+++ Editor/BuiltinPackage.cs
@@ -2,6 +2,8 @@
 using System;
 using System.Collections;
 using UnityEditor;
+using Hananoki.Extensions;
+using UnityEngine;
 
 using E = Hananoki.ManifestJsonUtility.SettingsEditor;
 
@@ -10,21 +12,23 @@ namespace Hananoki.ManifestJsonUtility {
 	public static class BuiltinPackage {
 		public static System.Collections.Hashtable dic;
 
-		public static bool GetDisplayName( out string displayName, string packageName ) {
+		public static bool GetDisplayName( ref PackageInfo info, string packageName ) {
 			if( dic == null ) {
 				dic = new Hashtable();
 				var files = DirectoryUtils.GetFiles( $"{EditorApplication.applicationContentsPath}/Resources/PackageManager/BuiltInPackages", "package.json", System.IO.SearchOption.AllDirectories );
 				foreach( var f in files ) {
 					var pj = PackageJson.Load( f );
 
-					dic.Add( pj.name, pj.displayName );
+					dic.Add( pj.name, pj );
 				}
 			}
 			if( dic.Contains( packageName ) ) {
-				displayName = (string) dic[ packageName ];
+				info.displayName = ((PackageJson) dic[ packageName ]).displayName;
+				info.version = ( (PackageJson) dic[ packageName ] ).version;
 				return true;
 			}
-			displayName = string.Empty;
+			info.displayName = string.Empty;
+			info.version = string.Empty;
 			return false;
 		}
 	}
@@ -33,21 +37,26 @@ namespace Hananoki.ManifestJsonUtility {
 	public static class PackageCache {
 		public static System.Collections.Hashtable dic;
 
-		public static bool GetDisplayName( out string displayName, string packageName ) {
+		public static bool GetDisplayName( ref PackageInfo info, string packageName ) {
 			if( dic == null ) {
 				dic = new Hashtable();
 				var files = DirectoryUtils.GetFiles( $"{Environment.CurrentDirectory}/Library/PackageCache", "package.json", System.IO.SearchOption.AllDirectories );
 				foreach( var f in files ) {
 					var pj = PackageJson.Load( f );
 
-					dic.Add( pj.name, pj.displayName );
+					dic.Add( pj.name, pj );
 				}
 			}
 			if( dic.Contains( packageName ) ) {
-				displayName = (string) dic[ packageName ];
+				info.displayName = ( (PackageJson) dic[ packageName ] ).displayName;
+				info.version = ( (PackageJson) dic[ packageName ] ).version;
+				if( info.version .StartsWith("http")) {
+					info.version = "URL";
+				}
 				return true;
 			}
-			displayName = string.Empty;
+			info.displayName = string.Empty;
+			info.version = string.Empty;
 			return false;
 		}
 	}
@@ -56,26 +65,32 @@ namespace Hananoki.ManifestJsonUtility {
 	public static class PackageUser {
 		public static System.Collections.Hashtable dic;
 
-		public static bool GetDisplayName( out string displayName, string packageName ) {
+		public static bool GetDisplayName( ref PackageInfo info, string packageName ) {
 			if( dic == null ) {
 				dic = new Hashtable();
 
 				E.Load();
 				foreach( var path in E.i.m_dirList ) {
+					if( path.IsEmpty() ) continue;
 					var files = DirectoryUtils.GetFiles( path, "package.json", System.IO.SearchOption.AllDirectories );
 					foreach( var fname in files ) {
-						//Debug.Log( sss );
 						var pj = PackageJson.Load( fname );
+						if( dic.Contains( pj.name ) ) {
+							Debug.LogWarning( $"Duplicate: {pj.name}: {fname}" );
+							continue;
+						}
 
-						dic.Add( pj.name, pj.displayName );
+						dic.Add( pj.name, pj );
 					}
 				}
 			}
 			if( dic.Contains( packageName ) ) {
-				displayName = (string) dic[ packageName ];
+				info.displayName = ( (PackageJson) dic[ packageName ] ).displayName;
+				info.version = ( (PackageJson) dic[ packageName ] ).version;
 				return true;
 			}
-			displayName = string.Empty;
+			info.displayName = string.Empty;
+			info.version = string.Empty;
 			return false;
 		}
 	}