+++ Editor/Types.cs
@@ -8,13 +8,15 @@ namespace Hananoki.ManifestJsonUtility {
 		public string displayName;
 		public Texture2D icon;
 		public PackageMode mode;
+		public string version;
 	}
 
 	public class PackageItem : TreeViewItem {
 		public string name;
-		public string version;
+		public string value;
 		public bool install;
 		public bool uninstall;
+		public string version;
 	}
 
 	public enum PackageMode {